+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
+### Added
+- Support the `X-HTTP-Method-Override` header.
+
 ### Fixed
 - No longer send a response body when one has not been supplied.